CS02 - SERS/src/requirements.txt
Flask
openai
python-dotenv
requests